Decoupled cache from nodes and moved to own file
